refactor: userProduct move to gift package
